# feat: Fused temperature online softmax kernel