--- a/csrc/flashinfer_sampling_ops.cu
+++ b/csrc/flashinfer_sampling_ops.cu
@@ -15,6 +15,8 @@
  */
 #include "pytorch_extension_utils.h"
 
+void softmax(at::Tensor logits, at::Tensor probs);
+
 void sampling_from_probs(at::Tensor probs, at::Tensor output,
                          std::optional<at::Tensor> maybe_indices, bool deterministic,
                          std::optional<at::Generator> gen);
@@ -60,6 +62,8 @@ void chain_speculative_sampling(at::Tensor draft_probs, at::Tensor draft_token_i
                                 std::optional<at::Generator> gen);
 
 TORCH_LIBRARY_FRAGMENT(TORCH_EXTENSION_NAME, m) {
+  // Softmax
+  m.def("softmax", softmax);
   // Sample from probabilities
   m.def("sampling_from_probs", sampling_from_probs);
   // Sample from logits