--- a/csrc/flashinfer_sampling_ops.cu
+++ b/csrc/flashinfer_sampling_ops.cu
@@ -16,7 +16,8 @@
 #include "pytorch_extension_utils.h"
 
 void softmax(at::Tensor workspace_buffer, at::Tensor logits, at::Tensor output,
-             std::optional<at::Tensor> maybe_temperature_arr, double temperature_val);
+             std::optional<at::Tensor> maybe_temperature_arr, double temperature_val,
+             bool enable_pdl);
 
 void sampling_from_probs(at::Tensor probs, at::Tensor output,
                          std::optional<at::Tensor> maybe_indices, bool deterministic,