… tests/test_sampling.py