@byjiang1996 2025-07-01
Hi @yzh119 @xslingcn , thanks for this awesome PR!

I tried to replace torch.div&torch.softmax to this flashinfer fuse kernel in sglang. [sglang code changes here](https://github.com/byjiang1996/sglang/commit/da5a65e76749e86cc8e74200fd5a08a55c2a7525)

However, in both cases, inference performance degrades after moving to the fuse kernel. Do you have any idea why?

## Before: via pytorch native
model sampler's div&softmax via pytorch native: 45us
<img width="998" alt="image" src="https://github.com/user-attachments/assets/324706c1-2601-4b4d-a834-dc367c09eefa" />

spec decoding verify's div&softmax via pytorch native: 28us
<img width="1220" alt="image" src="https://github.com/user-attachments/assets/29ed555f-dc5a-4c06-b7f7-fe073cf5e194" />

## After: via fuse kernel
model sampler's div&softmax via fuse kernel: 92us
<img width="680" alt="image" src="https://github.com/user-attachments/assets/7ecd7a90-bf18-4f92-a1a1-9d444b0d454d" />

spec decoding verify's div&softmax via fuse kernel: 69us
<img width="945" alt="image" src="https://github.com/user-attachments/assets/6ec96e2e-c4bc-46ed-8763-23d85783e45c" />